20230902: main page overhaul
this MAJOR update replaces our homepage with a SINoALICE-themed one in response to the Global Edition being shut down. this also includes metadata, typographical, and additional fixes and additions to enhance the experience.

also, `20230902` blogpost is getting some additions with response to the main page overhaul.

this event runs until 15 Nov 2023 at 6am UTC. afterwards, it'll be moved to the Main Page Archives.